app/src/main/java/com/example/mad_mp/HelperClass.java
package com.example.mad_mp;

public class HelperClass {
    String firstname, lastname, dateofbirth, username, email, password;
    String feedback;

    public String getFirstname() {
        return firstname;
    }

    public void setFirstname(String firstname) {
        this.firstname = firstname;
    }

    public String getLastname() {
        return lastname;
    }

    public void setLastname(String lastname) {
        this.lastname = lastname;
    }

    public String getDateofbirth() {
        return dateofbirth;
    }

    public void setDateofbirth(String dateofbirth) {
        this.dateofbirth = dateofbirth;
    }


    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }

    public String getFeedback() {
        return feedback;
    }
    public void setFeedback(String feedback) {
        this.feedback = feedback;
    }

    public HelperClass(String firstname, String lastname, String dateofbirth,
                       String username, String email, String password) {
        this.firstname = firstname;
        this.lastname = lastname;
        this.dateofbirth = dateofbirth;
        this.username = username;
        this.email = email;
        this.password = password;

    }


    public HelperClass( String feedback) {
        this.feedback = feedback;

    }

    // No-argument constructor
    public HelperClass() {
    }
}
